Added BlockSupport and TileEntitySupport. Made our packet system use an interface instead of an abstract class. Optimizations/bug fixes for BlockTrack/BlockDummy, TileEntityTrack, and the MultiBlock system. Tracks now save data, including MultiBlockStructure data, to NBT properly, so that they persist as they should. Added a new command and some other quality of life things for us to use whilst developing. The first of these is functionality to break tracks in case the MultiBlock system bugs out, if you get left over tile entities from tracks (the hitbox still being present, placing a block makes track reappear), hold the sneak button then break the block, this will bypass the MultiBlock system's destroyStructure() method.
The command, TestSupportRenderCommand, was added to allow for a way of changing the data in a TileEntitySupport whilst the TrackDesigner is being implemented. When first placed, BlockSupports are invisible, as none of the visibility data is set. To set the data, look at the block so that the game "selects" it (as if you were going to mine the block), and run the command "/support set" or "/s set", this will save the coordinates of that support block. Then run "/s [name of slot] [operation] <focus only>", where [name of slot] is defined in SupportUtils.SupportSlot (append "_base" or "_top" to modify those parts respectively), [operation] is any one of "true", "false", or "toggle", and <focus only> is a boolean that tells the command to only change the block at the coordinates that were set by "/s set". When <focus only> is false (its default), the command automatically checks the blocks immediately above and below the "focus" block (the coordinates set by "/s set"), and modifies their data accordingly, such that any supports with a block beneath them will show their base plate, supports with supports above and below will be just the pole, and the top-most support will show the top portion of the support pole. When <focus only> is true, only the "focus" block is modified.